requirements.txt
Flask
PyQt6
discord.py
pyinstaller